requirements.txt
asn1crypto==0.22.0
cffi==1.10.0
chardet==2.3.0
click==6.7
cryptography==1.8.1
enum34==1.1.6
Flask==0.12.1
future==0.16.0
idna==2.5
iotop==0.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
line-bot-sdk==1.1.0
lxml==3.5.0
MarkupSafe==1.0
packaging==16.8
pycparser==2.17
pymongo==3.4.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
uWSGI==2.0.15
Werkzeug==0.12.1
